Model/embeddings/requirements.txt
# Basic dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

# Sentence embedding dependencies
sentence-transformers>=2.2.0
torch>=1.9.0

# Word embedding dependencies (choose one)
# fasttext>=0.9.2
# gensim>=4.0.0

# Optional: for better performance
# huggingface_hub[hf_xet]>=0.20.0
